Move methods to a helper class
